+++ app/src/main/java/com/example/tournavigation/AppModule.kt
@@ -1,0 +1,7 @@
+package com.example.tournavigation
+
+import org.koin.dsl.module
+
+val appModel = module {
+    single { MapState(get())}
+}
