+++ app/src/main/java/com/example/tournavigation/MainApplication.kt
@@ -1,0 +1,14 @@
+package com.example.tournavigation
+
+import android.app.Application
+import org.koin.core.context.startKoin
+
+class MainApplication: Application() {
+        override fun onCreate() {
+            super.onCreate()
+
+            startKoin {
+                modules(appModel)
+            }
+        }
+}
