+++ app/src/main/java/com/example/tournavigation/MapState.kt
@@ -1,0 +1,7 @@
+package com.example.tournavigation
+
+import android.location.Location
+
+data class MapState(
+    val lastKnownLocation: Location?,
+)
