+++ build.gradle
@@ -1,6 +1,6 @@
 buildscript {
     ext {
-        compose_ui_version = '1.2.0'
+        compose_ui_version = '1.5.0'
     }
 }// Top-level build file where you can add configuration options common to all sub-projects/modules.
 plugins {
@@ -9,4 +9,8 @@ plugins {
     id 'org.jetbrains.kotlin.android' version '1.7.0' apply false
     id 'com.google.android.libraries.mapsplatform.secrets-gradle-plugin' version '2.0.1' apply false
     id 'org.jlleitschuh.gradle.ktlint' version '11.4.2'
+}
+
+task clean(type: Delete) {
+    delete rootProject.buildDir
 }
